chore: refact TestxServer methods and add method close
